Crate tokio_http2

HTTP/1.1: This library provides an HTTP library built on Futures and the Tokio Project for Async I/O. This version supports Pipelining for HTTP/1.1.

HTTP/2: (Interface being released soon) This library supplies the required modules to implement HTTP/2 which includes the HPACK header compression that includes the Huffman encoding/decoding features. This version will support Multiplexing which is required for HTTP/2.

Reexports

pub use status::StatusCode::{self, Ok, BadRequest, NotFound};

pub use version::HttpVersion;

pub use error::{Result, Error};

pub use method::Method;

pub use router::route::route::Route;

pub use router::Router;

pub use router::builder::RouterBuilder;

pub use logger::{Logger, LoggerLevel};

Modules

error
Error and Result module.

http

logger
Logger

method

router

server
The server-side abstraction for multipart requests. Enabled with the server feature (on by default).

status
HTTP status codes

version
HTTP Versions enum

Macros

chain_result

Structs

Request

Response

Url
A parsed URL record.

Type Definitions

Body

ContentLength

ContentType

Handler

Headers