Function tokio_http2::hpack::encoder::encode_integer

pub fn encode_integer(value: usize, prefix_size: u8) -> Vec<u8>

Encode an integer to the representation defined by HPACK.

Returns a newly allocated Vec containing the encoded bytes. Only prefix_size lowest-order bits of the first byte in the array are guaranteed to be used.