Module tokio_http2::http2

This library module provides HTTP/2 parsing and buffer frames used for HTTP/2.

NB: This code is changing so please do not depend on it at this time!

Modules

flag
NB: This code is changing so please do not depend on it at this time!

frame
NB: This code is changing so please do not depend on it at this time!

kind
NB: This code is changing so please do not depend on it at this time!

payload
NB: This code is changing so please do not depend on it at this time!

Structs

ErrorCode

ParserSettings

SizeIncrement

StreamIdentifier

Enums

Error
Errors that can occur during parsing an HTTP/2 frame.

HttpError

Constants

FRAME_HEADER_BYTES

Functions

encode_u24

encode_u32

encode_u64