Module tokio_http2::server

The server-side abstraction for multipart requests. Enabled with the server feature (on by default).

Use this when you are implementing an HTTP server and want to to accept, parse, and serve HTTP multipart/form-data requests (file uploads).

See the Multipart struct for more info.

Structs

Entries
A result of Multipart::save_all().

Multipart
The server-side implementation of multipart/form-data requests.

MultipartField
A field in a multipart request. May be either text or a binary stream (file).

MultipartFile
A representation of a file in HTTP multipart/form-data.

SavedFile
A file saved to the local filesystem from a multipart request.

Enums

MultipartData
The data of a field in a multipart/form-data request.

SaveDir
The save directory for Entries. May be temporary (delete-on-drop) or permanent.

SaveResult
The result of Multipart::save_all().

Traits

HttpRequest
A server-side HTTP request that may or may not be multipart.