Enum tokio_http2::server::MultipartData

pub enum MultipartData<'a, B: 'a> {
    Text(&'a str),
    File(MultipartFile<'a, B>),
}

The data of a field in a multipart/form-data request.

Variants
Text(&'a str)

The field's payload is a text string.

File(MultipartFile<'a, B>)

The field's payload is a binary stream (file).

Methods

impl<'a, B> MultipartData<'a, B>

fn as_text(&self) -> Option<&str>

Borrow this payload as a text field, if possible.

fn as_file(&mut self) -> Option<&mut MultipartFile<'a, B>>

Borrow this payload as a file field, if possible. Mutably borrows so the contents can be read.