Struct tokio_http2::server::SavedFile

pub struct SavedFile {
    pub path: PathBuf,
    pub filename: Option<String>,
    pub size: u64,
}

A file saved to the local filesystem from a multipart request.

Fields
path: PathBuf

The complete path this file was saved at.

filename: Option<String>

The original filename of this file, if one was provided in the request.

Warning

You should treat this value as untrustworthy because it is an arbitrary string provided by the client. You should not blindly append it to a directory path and save the file there, as such behavior could easily be exploited by a malicious client.

size: u64

The number of bytes written to the disk; may be truncated.